requirements.txt
google-generativeai==0.8.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
